Add suggestions
When the first page of the wikipedia search does not contain a ".bday"
element, we display the list of the other results as suggestions for the
right page.

For instance, searching for "prince" will return the page about actual
princes, but the second page would be "prince (musician)", so that's
what we want.

The form is not auto-submitted, only the input is auto-filled, because
triggering a form submission that will trigger event listeners is a pain
with vanilla javascript.